Initial split of intrin/ and arch/.
