+++ src/arch/mod.rs
@@ -1,0 +1,2 @@
+mod x86;
+
