+++ src/arch/x86/intrin/swizzle.rs
