+++ src/into_iters.rs
@@ -2,7 +2,7 @@
 // Copyright 2017 Adam Niederer <adam.niederer@gmail.com>
 
 // This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
+// License, v. 2.0. If a copy owf the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 use crate::iters::{SIMDIter, SIMDIterator, SIMDObject};
