+++ src/intrin/mod.rs
@@ -1,60 +1,52 @@
-// This file is part of faster, the SIMD library for humans.
-// Copyright 2017 Adam Niederer <adam.niederer@gmail.com>
+pub mod abs;
+pub mod addsub;
+pub mod cast;
+pub mod cmp;
+#[macro_use] pub mod destride;
+pub mod downcast;
+#[macro_use] pub mod endian;
+#[macro_use] pub mod eq;
+pub mod hadd;
+pub mod hsub;
+#[macro_use] pub mod macros;
+#[macro_use] pub mod merge;
+#[macro_use] pub mod popcnt;
+pub mod recip;
+pub mod round;
+pub mod rsqrt;
+#[macro_use] pub mod sum;
+pub mod saturating_add;
+pub mod saturating_hadd;
+pub mod saturating_hsub;
+pub mod saturating_sub;
+pub mod sqrt;
+#[macro_use] pub mod transmute;
+pub mod upcast;
 
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#![allow(unused_imports)]
-
-#[macro_use]
-mod macros;
-
-mod abs;
-pub use self::abs::*;
-mod sqrt;
-pub use self::sqrt::*;
-mod transmute;
-pub use self::transmute::*;
-mod cast;
-pub use self::cast::*;
-mod upcast;
-pub use self::upcast::*;
-mod downcast;
-pub use self::downcast::*;
-mod round;
-pub use self::round::*;
-mod recip;
-pub use self::recip::*;
-mod hadd;
-pub use self::hadd::*;
-mod hsub;
-pub use self::hsub::*;
-mod rsqrt;
-pub use self::rsqrt::*;
-mod cmp;
-pub use self::cmp::*;
-mod eq;
-pub use self::eq::*;
-mod saturating_add;
-pub use self::saturating_add::*;
-mod saturating_hadd;
-pub use self::saturating_hadd::*;
-mod saturating_sub;
-pub use self::saturating_sub::*;
-mod saturating_hsub;
-pub use self::saturating_hsub::*;
-mod addsub;
-pub use self::addsub::*;
-mod merge;
-pub use self::merge::*;
-mod swizzle;
-pub use self::swizzle::*;
-mod endian;
-pub use self::endian::*;
-mod sum;
-pub use self::sum::*;
-mod popcnt;
-pub use self::popcnt::*;
-mod destride;
-pub use self::destride::*;
+// We use an internal prelude not to clutter the namespace when we import
+// from actual prelude. 
+pub(crate) mod prelude {
+    pub use super::abs::*;
+    pub use super::addsub::*;
+    pub use super::cast::*;
+    pub use super::cmp::*;
+    pub use super::destride::*;
+    pub use super::downcast::*;
+    pub use super::endian::*;
+    pub use super::eq::*;
+    pub use super::hadd::*;
+    pub use super::hsub::*;
+    pub use super::merge::*;
+    pub use super::popcnt::*;
+    pub use super::recip::*;
+    pub use super::round::*;
+    pub use super::rsqrt::*;
+    pub use super::sum::*;
+    pub use super::saturating_add::*;
+    pub use super::saturating_hadd::*;
+    pub use super::saturating_hsub::*;
+    pub use super::saturating_sub::*;
+    pub use super::transmute::*;
+    pub use super::upcast::*;
+}
