+++ src/prelude.rs
@@ -9,7 +9,7 @@ pub use crate::iters::*;
 pub use crate::into_iters::*;
 pub use crate::vecs::{Packed, Pattern};
 pub use crate::vecs::{u8s, i8s, u16s, i16s, u32s, i32s, f32s, u64s, i64s, f64s};
-pub use crate::intrin::*;
+pub use crate::intrin::prelude::*;
 pub use crate::zip::*;
 pub use crate::stride_zip::*;
 pub use crate::stride::*;
